tests/robustness: Extract validation to separate package
